requirements.txt
Django==6.0.3
environs[django]==14.*
pillow==12.1.1
djangorestframework==3.17.1